pyproject.toml
[tool.poetry]
name = "parallax-ops"
version = "0.1.0"
description = "Terminal-based AI copilot for DevOps engineers"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.0"
openai = "^1.0.0"
pydantic = "^2.0.0"
pyyaml = "^6.0"
pyperclip = "^1.8.0"

[tool.poetry.scripts]
pop = "src.main:main"

[project]
name = "parallax-ops"
version = "0.1.0"
description = "Terminal-based AI copilot for DevOps engineers"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "pyperclip>=1.8.0",
]

[project.scripts]
pop = "src.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
